+++ sickrage/metadata/helpers.py
@@ -180,7 +180,11 @@ def qualityFromFileMeta(filename):
         bluray = re.search(r"blue?-?ray|hddvd|b[rd](rip|mux)", base_filename, re.I) is not None
         webdl = re.search(r"web.?dl|web(rip|mux|hd)", base_filename, re.I) is not None
 
-        if data['resolution_height'] > 1000:
+        if 3240 < data['resolution_height']:
+            ret = ((Quality.UHD_8K_TV, Quality.UHD_8K_BLURAY)[bluray], Quality.UHD_8K_WEBDL)[webdl]
+        if 1620 < data['resolution_height'] <= 3240:
+            ret = ((Quality.UHD_4K_TV, Quality.UHD_4K_BLURAY)[bluray], Quality.UHD_4K_WEBDL)[webdl]
+        elif 800 < data['resolution_height'] <= 1620:
             quality = ((Quality.FULLHDTV, Quality.FULLHDBLURAY)[bluray], Quality.FULLHDWEBDL)[webdl]
         elif 680 < data['resolution_height'] < 800:
             quality = ((Quality.HDTV, Quality.HDBLURAY)[bluray], Quality.HDWEBDL)[webdl]